Bugfix SDO forwarding: "Default OBD access function not call if object is defined"
All mappable objects have to be defined at AP side.
If a mappable object is accessed via SDO it will not be forwarded to the AP if it is defined.
Thus, the default OBD access function has to be called also, if and object >= 0x2000 has been accessed -
not only if it does not exist.